feat: Added fetchMore on albums view
